+++ src/coinbase/address/wallet_address.ts
@@ -148,6 +148,7 @@ export class WalletAddress extends Address {
    * @param options.destination - The destination of the transfer. If a Wallet, sends to the Wallet's default address. If a String, interprets it as the address ID.
    * @param options.timeoutSeconds - The maximum amount of time to wait for the Transfer to complete, in seconds.
    * @param options.intervalSeconds - The interval at which to poll the Network for Transfer status, in seconds.
+   * @param options.gasless - Whether the Transfer should be gasless. Defaults to false.
    * @returns The transfer object.
    * @throws {APIError} if the API request to create a Transfer fails.
    * @throws {APIError} if the API request to broadcast a Transfer fails.
@@ -159,6 +160,7 @@ export class WalletAddress extends Address {
     destination,
     timeoutSeconds = 10,
     intervalSeconds = 0.2,
+    gasless = false,
   }: CreateTransferOptions): Promise<Transfer> {
     if (!Coinbase.useServerSigner && !this.key) {
       throw new InternalError("Cannot transfer from address without private key loaded");
@@ -180,6 +182,7 @@ export class WalletAddress extends Address {
       network_id: destinationNetworkId,
       asset_id: asset.primaryDenomination(),
       destination: destinationAddress,
+      gasless: gasless,
     };
 
     let response = await Coinbase.apiClients.transfer!.createTransfer(
@@ -192,22 +195,9 @@ export class WalletAddress extends Address {
 
     if (!Coinbase.useServerSigner) {
       const wallet = new ethers.Wallet(this.key!.privateKey);
-      const transaction = transfer.getTransaction();
-      let signedPayload = await wallet!.signTransaction(transaction);
-      signedPayload = signedPayload.slice(2);
+      await transfer.sign(wallet);
 
-      const broadcastTransferRequest = {
-        signed_payload: signedPayload,
-      };
-
-      response = await Coinbase.apiClients.transfer!.broadcastTransfer(
-        this.getWalletId(),
-        this.getId(),
-        transfer.getId(),
-        broadcastTransferRequest,
-      );
-
-      transfer = Transfer.fromModel(response.data);
+      transfer = await transfer.broadcast();
     }
 
     const startTime = Date.now();