+++ src/tests/utils.ts
@@ -14,6 +14,7 @@ import {
   ValidatorList,
   Validator,
   StakingOperationStatusEnum,
+  FeatureSet,
 } from "../client";
 import { BASE_PATH } from "../client/base";
 import { Coinbase } from "../coinbase/coinbase";
@@ -98,7 +99,7 @@ export const VALID_ADDRESS_MODEL = newAddressModel(randomUUID());
 export const VALID_WALLET_MODEL: WalletModel = {
   id: randomUUID(),
   network_id: Coinbase.networks.BaseSepolia,
-  enabled_features: [],
+  feature_set: {} as FeatureSet,
   default_address: {
     wallet_id: walletId,
     address_id: "0xdeadbeef",
@@ -139,6 +140,32 @@ export const VALID_TRANSFER_MODEL: TransferModel = {
   destination: "0x4D9E4F3f4D1A8B5F4f7b1F5b5C7b8d6b2B3b1b0b",
   asset_id: Coinbase.assets.Eth,
   amount: new Decimal(ethers.parseUnits("100", 18).toString()).toString(),
+  gasless: false,
+};
+
+export const VALID_TRANSFER_SPONSORED_SEND_MODEL: TransferModel = {
+  transfer_id: transferId,
+  network_id: Coinbase.networks.BaseSepolia,
+  wallet_id: walletId,
+  asset: {
+    asset_id: Coinbase.assets.Usdc,
+    network_id: Coinbase.networks.BaseSepolia,
+    decimals: 18,
+    contract_address: "0xusdc",
+  },
+  sponsored_send: {
+    to_address_id: "0xdeadbeef",
+    raw_typed_data: "0xhash",
+    typed_data_hash: "0x7523946e17c0b8090ee18c84d6f9a8d63bab4d579a6507f0998dde0791891823",
+    transaction_hash: "0xdeadbeef",
+    transaction_link: "https://sepolia.basescan.org/tx/0xdeadbeef",
+    status: "pending",
+  },
+  address_id: ethers.Wallet.createRandom().address,
+  destination: "0x4D9E4F3f4D1A8B5F4f7b1F5b5C7b8d6b2B3b1b0b",
+  asset_id: Coinbase.assets.Eth,
+  amount: new Decimal(ethers.parseUnits("100", 18).toString()).toString(),
+  gasless: false,
 };
 
 export const VALID_STAKING_OPERATION_MODEL: StakingOperationModel = {
@@ -419,6 +446,7 @@ export const stakeApiMock = {
   getExternalStakingOperation: jest.fn(),
   getStakingContext: jest.fn(),
   fetchStakingRewards: jest.fn(),
+  fetchHistoricalStakingBalances: jest.fn(),
   broadcastStakingOperation: jest.fn(),
   createStakingOperation: jest.fn(),
   getStakingOperation: jest.fn(),
@@ -433,6 +461,7 @@ export const externalAddressApiMock = {
   listExternalAddressBalances: jest.fn(),
   getExternalAddressBalance: jest.fn(),
   requestExternalFaucetFunds: jest.fn(),
+  listAddressHistoricalBalance: jest.fn(),
 };
 
 export const serverSignersApiMock = {