+++ src/tests/wallet_address_test.ts
@@ -6,6 +6,7 @@ import { FaucetTransaction } from "../coinbase/faucet_transaction";
 import {
   Balance as BalanceModel,
   FetchStakingRewards200Response,
+  FetchHistoricalStakingBalances200Response,
   StakingContext as StakingContextModel,
   StakingOperation as StakingOperationModel,
   StakingOperationStatusEnum,
@@ -46,6 +47,7 @@ import { WalletAddress } from "../coinbase/address/wallet_address";
 import { Wallet } from "../coinbase/wallet";
 import { StakingOperation } from "../coinbase/staking_operation";
 import { StakingReward } from "../coinbase/staking_reward";
+import { StakingBalance } from "../coinbase/staking_balance";
 
 // Test suite for the WalletAddress class
 describe("WalletAddress", () => {
@@ -309,6 +311,11 @@ describe("WalletAddress", () => {
           amount: "361",
           state: StakingRewardStateEnum.Pending,
           format: StakingRewardFormat.Usd,
+          usd_value: {
+            amount: "361",
+            conversion_price: "3000",
+            conversion_time: "2024-05-01T00:00:00Z",
+          },
         },
         {
           address_id: newAddress.address_id,
@@ -316,6 +323,11 @@ describe("WalletAddress", () => {
           amount: "203",
           state: StakingRewardStateEnum.Pending,
           format: StakingRewardFormat.Usd,
+          usd_value: {
+            amount: "203",
+            conversion_price: "3000",
+            conversion_time: "2024-05-02T00:00:00Z",
+          },
         },
         {
           address_id: newAddress.address_id,
@@ -323,6 +335,60 @@ describe("WalletAddress", () => {
           amount: "226",
           state: StakingRewardStateEnum.Pending,
           format: StakingRewardFormat.Usd,
+          usd_value: {
+            amount: "226",
+            conversion_price: "3000",
+            conversion_time: "2024-05-03T00:00:00Z",
+          },
+        },
+      ],
+      has_more: false,
+      next_page: "",
+    };
+
+    const HISTORICAL_STAKING_BALANCES_RESPONSE: FetchHistoricalStakingBalances200Response = {
+      data: [
+        {
+          address: newAddress.address_id,
+          date: "2024-05-01",
+          bonded_stake: {
+            amount: "32",
+            asset: {
+              asset_id: Coinbase.assets.Eth,
+              network_id: Coinbase.networks.EthereumHolesky,
+              decimals: 18,
+            },
+          },
+          unbonded_balance: {
+            amount: "2",
+            asset: {
+              asset_id: Coinbase.assets.Eth,
+              network_id: Coinbase.networks.EthereumHolesky,
+              decimals: 18,
+            },
+          },
+          participant_type: "validator",
+        },
+        {
+          address: newAddress.address_id,
+          date: "2024-05-02",
+          bonded_stake: {
+            amount: "34",
+            asset: {
+              asset_id: Coinbase.assets.Eth,
+              network_id: Coinbase.networks.EthereumHolesky,
+              decimals: 18,
+            },
+          },
+          unbonded_balance: {
+            amount: "3",
+            asset: {
+              asset_id: Coinbase.assets.Eth,
+              network_id: Coinbase.networks.EthereumHolesky,
+              decimals: 18,
+            },
+          },
+          participant_type: "validator",
         },
       ],
       has_more: false,
@@ -462,6 +528,30 @@ describe("WalletAddress", () => {
         Coinbase.apiClients.asset!.getAsset = getAssetMock();
         const response = await walletAddress.stakingRewards(Coinbase.assets.Eth);
         expect(response).toBeInstanceOf(Array<StakingReward>);
+      });
+    });
+
+    describe(".historicalStakingBalances", () => {
+      it("should successfully return historical staking balances", async () => {
+        Coinbase.apiClients.stake!.fetchHistoricalStakingBalances = mockReturnValue(
+          HISTORICAL_STAKING_BALANCES_RESPONSE,
+        );
+        Coinbase.apiClients.asset!.getAsset = getAssetMock();
+        const response = await walletAddress.historicalStakingBalances(Coinbase.assets.Eth);
+        expect(response).toBeInstanceOf(Array<StakingBalance>);
+        expect(response.length).toEqual(2);
+        expect(response[0].bondedStake().amount).toEqual(new Decimal("32"));
+        expect(response[0].bondedStake().asset?.assetId).toEqual(Coinbase.assets.Eth);
+        expect(response[0].bondedStake().asset?.decimals).toEqual(18);
+        expect(response[0].bondedStake().asset?.networkId).toEqual(
+          Coinbase.networks.EthereumHolesky,
+        );
+        expect(response[0].unbondedBalance().amount).toEqual(new Decimal("2"));
+        expect(response[0].unbondedBalance().asset?.assetId).toEqual(Coinbase.assets.Eth);
+        expect(response[0].unbondedBalance().asset?.decimals).toEqual(18);
+        expect(response[0].unbondedBalance().asset?.networkId).toEqual(
+          Coinbase.networks.EthereumHolesky,
+        );
       });
     });
   });